+++ src/modules/getFeatureInfo/components/GetFeatureInfoDetached.vue
@@ -45,6 +45,7 @@ export default {
             "centerMapToClickPoint",
             "currentFeature",
             "highlightVectorRules",
+            "showPolygonMarkerForWMS",
             "menuSide",
             "showMarker",
             "hideMapMarkerOnVectorHighlight"
@@ -74,11 +75,13 @@ export default {
     },
     mounted () {
         this.highlightVectorFeature();
+        this.highlightWMSFeature();
         this.setMarker();
     },
     updated: function () {
         if (this.isUpdated) {
             this.highlightVectorFeature();
+            this.highlightWMSFeature();
             this.setMarker();
             this.$emit("updateFeatureDone");
         }
@@ -87,6 +90,7 @@ export default {
         if (this.searchInput === "") {
             this.removePointMarker();
             this.removeHighlighting();
+            this.removePolygonMarker();
         }
     },
     methods: {
@@ -94,6 +98,7 @@ export default {
         ...mapActions("Maps", [
             "placingPointMarker",
             "removePointMarker",
+            "removePolygonMarker",
             "highlightFeature",
             "removeHighlightFeature",
             "setCenter"
@@ -178,6 +183,14 @@ export default {
                             };
                             break;
                         }
+                        case "MultiLineString":
+                        {
+                            highlightObject.type = "highlightMultiLine";
+                            highlightObject.highlightStyle = {
+                                stroke: this.highlightVectorRules.stroke
+                            };
+                            break;
+                        }
                         default:
                             break;
                     }
@@ -186,6 +199,62 @@ export default {
                     this.highlightFeature(highlightObject);
                 }
                 this.lastFeature = this.feature;
+            }
+        },
+        /**
+         * Highlights a feature with a Polygon-Marker
+         * @returns {void}
+         */
+        highlightWMSFeature () {
+            if (this.showPolygonMarkerForWMS) {
+                const layer = layerCollection.getLayerById(this.feature.getLayerId()),
+                    highlightObject = {
+                        feature: this.feature.getOlFeature(),
+                        layer: {id: this.feature.getLayerId()}
+                    };
+
+                if (layer?.attributes?.typ?.toLowerCase() === "wms") {
+                    this.removePolygonMarker();
+
+                    if (this.hideMapMarkerOnVectorHighlight) {
+                        this.hideMarker();
+                    }
+
+                    if (this.feature.getOlFeature() && typeof this.feature.getOlFeature().getGeometry === "function") {
+                        switch (this.feature.getOlFeature().getGeometry()?.getType()) {
+                            case "Point": {
+                                highlightObject.type = "highlightPoint";
+                                break;
+                            }
+                            case "MultiPoint": {
+                                highlightObject.type = "highlightMultiPoint";
+                                break;
+                            }
+                            case "Polygon": {
+                                highlightObject.type = "highlightPolygon";
+                                break;
+                            }
+                            case "MultiPolygon": {
+                                highlightObject.type = "highlightMultiPolygon";
+                                break;
+                            }
+                            case "LineString": {
+                                highlightObject.type = "highlightLine";
+                                break;
+                            }
+                            case "MultiLineString": {
+                                highlightObject.type = "highlightMultiLine";
+                                break;
+                            }
+                            default:
+                                break;
+                        }
+                    }
+                    if (highlightObject.type) {
+                        this.highlightFeature(highlightObject);
+                    }
+                    this.lastFeature = this.feature;
+                }
             }
         },
         /**